requirements.txt
certifi==2017.7.27.1
chardet==3.0.4
coreapi==2.3.0
coreschema==0.0.4
Django==2.1.3
django-cors-headers==2.0.2
djangorestframework==3.9.0
idna==2.6
itypes==1.1.0
Jinja2==2.9.6
Markdown==2.6.8
MarkupSafe==1.0
olefile==0.44
Pillow==4.1.1
psycopg2==2.7.3.2
Pygments==2.2.0
pytz==2017.3
requests==2.18.4
sqlparse==0.2.4
uritemplate==3.0.0
urllib3==1.22
djangorestframework-jwt==1.11.0
django-debug-toolbar==1.10.1